I am doing.

+++ handlers/users/menu/menu.py
@@ -48,40 +48,84 @@ Menu Product Info and modifications list
 """
 
 
-#
 # @dp.callback_query_handler(state=MenuState.product)
 # async def choose_product(call: types.CallbackQuery, state: FSMContext):
 #     if call.data != 'back_to_categories':
-#         product = afc.get_product(call.data)
-#         await call.message.delete()
 #         async with state.proxy() as data:
 #             data['quantity']: int = 1
 #             data['product_id'] = call.data
-#             try:
-#                 try:
-#                     await MenuState.modifications.set()
-#                     await call.message.answer_photo(photo=afc.base_url + product['photo'] + '?token=' + afc.api_key,
-#                                                     caption=f"{product['product_name']}\n",
-#                                                     reply_markup=choose_modification_key(
-#                                                         modifications=product['modifications']))
-#                 except:
-#                     await MenuState.modifications.set()
-#                     await call.message.answer_photo(photo=afc.base_url + product['photo'] + '?token=' + afc.api_key,
-#                                                     caption=f"{product['product_name']}\n",
-#                                                     reply_markup=dish_modification_key(
-#                                                         product['group_modifications'][0]['modifications']))
-#             except:
+#             product = afc.get_product(call.data)
+#             if 'modifications' in product:
+#                 await call.message.answer_photo(photo=afc.base_url + product['photo'],
+#                                                 caption=f"{product['product_name']}\n",
+#                                                 reply_markup=choose_modification_key(
+#                                                     modifications=product['modifications'],
+#                                                     type='modifications'))
+#                 await call.message.delete()
+#                 await MenuState.modifications.set()
+#             elif 'group_modifications' in product:
+#                 if len(product['group_modifications']) == 1:
+#                     if product['group_modifications'][0]['type'] == 1:
+#                         await call.message.answer_photo(photo=afc.base_url + product['photo'],
+#                                                         caption=f"{product['product_name']}\n",
+#                                                         reply_markup=choose_modification_key(
+#                                                             modifications=product['group_modifications'][0][
+#                                                                 'modifications'],
+#                                                             type='group_modifications'
+#                                                         ))
+#                         await call.message.delete()
+#                         await MenuState.modifications.set()
+#                     else:
+#                         if product['group_modifications'][0]['num_min'] != product['group_modifications'][0]['num_max']:
+#                             await call.message.answer_photo(photo=afc.base_url + product['photo'],
+#                                                             caption=f"{product['product_name']}\n\n\n",
+#                                                             reply_markup=choose_modification_key(
+#                                                                 modifications=product['group_modifications'][0][
+#                                                                     'modifications'],
+#                                                                 type='group_modifications'
+#                                                             ))
+#                             await call.message.delete()
+#                             await MenuState.modifications.set()
+#                         else:
+#                                 caption = ''
+#                                 caption += f"{product['product_name']}\n"
+#                                 caption += f"Narxi: <b>{product['sources'][1]['price'][:-2]}</b>\n\n\n"
+#                                 caption += f"Combo Quyidagilarni o'z ichiga oladi:\n\n"
+#                                 count = 1
+#                                 for combo_item in product['group_modifications'][0]['modifications']:
+#                                     caption += f"{count}.{combo_item['name']} - {combo_item['brutto']}ta\n"
+#                                     count += 1
+#                                     await call.message.answer_photo(photo=afc.base_url + product['photo'],
+#                                                                 caption=caption,
+#                                                                 reply_markup=booking_key(data['quantity']))
+#                                     await call.message.delete()
+#                                     await MenuState.booking.set()
+
+
+#                 else:
+#                     caption = ''
+#                     caption += f"{product['product_name']}\n"
+#                     caption += f"Narxi: <b>{product['sources'][1]['price'][:-2]}</b>\n\n\n"
+#                     caption += f"Combo Quyidagilarni o'z ichiga oladi:\n\n"
+#                     count = 1
+#                     for combo_item in product['group_modifications']:
+#                         caption += f"{count}.{combo_item['modifications'][0]['name']} - {combo_item['modifications'][0]['brutto']}ta\n"
+#                         count += 1
+#                     await call.message.answer_photo(photo=afc.base_url + product['photo'],
+#                                                     caption=caption,
+#                                                     reply_markup=booking_key(data['quantity']))
+#                     await call.message.delete()
+#                     await MenuState.booking.set()
+#             else:
 #                 await call.message.answer_photo(photo=afc.base_url + product['photo'] + '?token=' + afc.api_key,
 #                                                 caption=f"{product['product_name']}\n"
 #                                                 # f"{product['product_production_description']}\n"
 #                                                         f"{product['spots'][0]['price'][:-2]} so'm",
 #                                                 reply_markup=booking_key(data['quantity']))
 #                 await MenuState.booking.set()
-#
-#
 #     else:
+#         await state.finish()
 #         await menu(call)
-
 
 @dp.callback_query_handler(state=MenuState.product)
 async def choose_product(call: types.CallbackQuery, state: FSMContext):
@@ -95,9 +139,11 @@ async def choose_product(call: types.CallbackQuery, state: FSMContext):
                                                 caption=f"{product['product_name']}\n",
                                                 reply_markup=choose_modification_key(
                                                     modifications=product['modifications'],
-                                                    type='modifications'))
+                                                    type='modifications'
+                                                ))
                 await call.message.delete()
                 await MenuState.modifications.set()
+
             elif 'group_modifications' in product:
                 if len(product['group_modifications']) == 1:
                     if product['group_modifications'][0]['type'] == 1:
@@ -110,6 +156,7 @@ async def choose_product(call: types.CallbackQuery, state: FSMContext):
                                                         ))
                         await call.message.delete()
                         await MenuState.modifications.set()
+
                     else:
                         if product['group_modifications'][0]['num_min'] != product['group_modifications'][0]['num_max']:
                             await call.message.answer_photo(photo=afc.base_url + product['photo'],
@@ -121,6 +168,7 @@ async def choose_product(call: types.CallbackQuery, state: FSMContext):
                                                             ))
                             await call.message.delete()
                             await MenuState.modifications.set()
+
                         else:
                             caption = ''
                             caption += f"{product['product_name']}\n"
@@ -135,30 +183,34 @@ async def choose_product(call: types.CallbackQuery, state: FSMContext):
                                                             reply_markup=booking_key(data['quantity']))
                             await call.message.delete()
                             await MenuState.booking.set()
-                else:
+
+
+
+                elif len(product['group_modifications']) > 1:
+                    print(f"Product : - {product}")
                     caption = ''
                     caption += f"{product['product_name']}\n"
-                    caption += f"Narxi: <b>{product['sources'][1]['price'][:-2]}</b>\n\n\n"
+                    caption += f"Narxi: <b>{product['sources'][1]['price'][:-2]}</b>\n\n\n so'm"
                     caption += f"Combo Quyidagilarni o'z ichiga oladi:\n\n"
                     count = 1
                     for combo_item in product['group_modifications']:
-                        caption += f"{count}.{combo_item['modifications'][0]['name']} - {combo_item['modifications'][0]['brutto']}ta\n"
+                        print(combo_item)
+
+                        caption += f"{count}.{combo_item['modifications'][0]['name']} - {combo_item['modifications'][0]['brutto']} ta\n"
                         count += 1
                     await call.message.answer_photo(photo=afc.base_url + product['photo'],
                                                     caption=caption,
                                                     reply_markup=booking_key(data['quantity']))
                     await call.message.delete()
                     await MenuState.booking.set()
+
             else:
-                await call.message.answer_photo(photo=afc.base_url + product['photo'] + '?token=' + afc.api_key,
+                await call.message.answer_photo(photo=afc.base_url + product['photo'],
                                                 caption=f"{product['product_name']}\n"
                                                 # f"{product['product_production_description']}\n"
                                                         f"{product['spots'][0]['price'][:-2]} so'm",
                                                 reply_markup=booking_key(data['quantity']))
                 await MenuState.booking.set()
-    else:
-        await state.finish()
-        await menu(call)
 
 
 """
@@ -177,13 +229,16 @@ async def choose_modification(call: types.CallbackQuery, state: FSMContext):
             price = ''
             if 'modifications' in product:
                 for modification in product['modifications']:
+
                     if modification['modificator_id'] == call.data:
                         price = modification['sources'][1]['price'][:-2]
                         break
             elif "group_modifications" in product:
+                print(product)
                 for modification in product['group_modifications'][0]['modifications']:
                     if modification['dish_modification_id'] == int(call.data):
                         price = modification['price']
+
                         break
             await call.message.edit_caption(caption=f"{product['product_name']}\n"
                                                     f"{price} so'm",
